fix(python): harden cargo allowlist, cross-platform tests, split test file
- Add _CARGO_ALLOWED frozenset to restrict cargo subcommands in run()
- Replace Unix-only commands in test_platform.py with sys.executable -c equivalents
- Split test_connectors.py (644 lines) into test_connectors_github.py and test_connectors_package_manager.py
- Fix duplicate 'client' key in pyproject.toml optional-dependencies
- Add test for cargo unsafe subcommand rejection
